[EC-428] Moved ICreateProviderCommand to Core project
