--- src/Admin/Providers/Interfaces/ICreateProviderCommand.cs
@@ -1,9 +1,0 @@
-﻿using Bit.Admin.Models;
-using Bit.Core.Entities.Provider;
-
-namespace Bit.Admin.Providers.Interfaces;
-
-public interface ICreateProviderCommand
-{
-    Task<Provider> Create(CreateProviderModel model);
-}
