+++ src/Core/AdminFeatures/Providers/Interfaces/ICreateProviderCommand.cs
@@ -1,0 +1,9 @@
+﻿using Bit.Core.Entities.Provider;
+
+namespace Bit.Core.AdminFeatures.Providers.Interfaces;
+
+public interface ICreateProviderCommand
+{
+    Task CreateMspAsync(Provider provider, string ownerEmail);
+    Task CreateResellerAsync(Provider provider);
+}
