+++ src/SharedWeb/Utilities/ServiceCollectionExtensions.cs
@@ -2,6 +2,7 @@
 using System.Security.Claims;
 using System.Security.Cryptography.X509Certificates;
 using AspNetCoreRateLimit;
+using Bit.Core.AdminFeatures;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.HostedServices;
@@ -111,6 +112,7 @@ public static class ServiceCollectionExtensions
 
     public static void AddBaseServices(this IServiceCollection services, IGlobalSettings globalSettings)
     {
+        services.AddProviderCommands();
         services.AddScoped<ICipherService, CipherService>();
         services.AddScoped<IUserService, UserService>();
         services.AddOrganizationServices(globalSettings);
